chore(contentful): move contents and search to different folders
