+++ packages/botonic-plugin-contentful/tests/contentful/contents/carousel.test.ts
@@ -1,8 +1,9 @@
-import { testContentful, testContext } from './contentful.helper'
-import * as cms from '../../src'
+import { testContentful, testContext } from '../contentful.helper'
+import * as cms from '../../../src'
 import { expectImgUrlIs } from './image.test'
 
 export const TEST_CAROUSEL_MAIN_ID = '2yR9f3stNAEqdamUr8VtfD'
+export const TEST_POST_MENU_CRSL = '7ATP377r6Nb8VqNtRKAS4J'
 
 function assertElementDudasPrevias(element: cms.Element) {
   expect(element.title).toBe('Dudas previas a la compra')
