+++ packages/botonic-plugin-contentful/tests/contentful/contents/date-range.test.ts
@@ -1,4 +1,4 @@
-import { testContentful } from './contentful.helper'
+import { testContentful } from '../contentful.helper'
 
 const TEST_DATE_RANGE_HUBTYPE_ID = '46taK0TceR2OZ8issCMhB7'
