+++ packages/botonic-plugin-contentful/tests/contentful/contents/text.test.ts
@@ -1,6 +1,6 @@
 import { expectImgUrlIs } from './image.test'
-import { testContentful, testContext } from './contentful.helper'
-import * as cms from '../../src'
+import { testContentful, testContext } from '../contentful.helper'
+import * as cms from '../../../src'
 
 export const TEST_POST_FAQ1_ID = 'djwHOFKknJ3AmyG6YKNip'
 const TEST_POST_FAQ2_ID = '22h2Vba7v92MadcL5HeMrt'
