+++ packages/botonic-plugin-contentful/tests/contentful/keywords.test.ts
@@ -1,7 +1,7 @@
 import { ContentCallback, ContentType } from '../../src/cms'
 import { SearchResult } from '../../src/search'
 import { testContentful, testContext } from './contentful.helper'
-import { TEST_POST_FAQ1_ID } from './text.test'
+import { TEST_POST_FAQ1_ID } from './contents/text.test'
 import 'jest-extended'
 
 test('TEST: contentful contentsWithKeywords', async () => {
